fix: remove usize in AMT
1. Use u32 for bitwidth/height.
2. Use u64 everywhere else.

+++ ipld/amt/src/lib.rs
@@ -18,28 +18,28 @@ pub(crate) use self::node::Node;
 pub(crate) use self::root::Root;
 pub use self::value_mut::ValueMut;
 
-const DEFAULT_BIT_WIDTH: usize = 3;
-const MAX_HEIGHT: usize = 64;
+const DEFAULT_BIT_WIDTH: u32 = 3;
+const MAX_HEIGHT: u32 = 64;
 
 /// MaxIndex is the maximum index for elements in the AMT. This u64::MAX-1 so we
 /// don't overflow u64::MAX when computing the length.
-pub const MAX_INDEX: usize = (std::u64::MAX - 1) as usize;
+pub const MAX_INDEX: u64 = (std::u64::MAX - 1) as u64;
 
-fn nodes_for_height(bit_width: usize, height: usize) -> usize {
-    let height_log_two = bit_width * height;
+fn nodes_for_height(bit_width: u32, height: u32) -> u64 {
+    let height_log_two = bit_width as u64 * height as u64;
     if height_log_two >= 64 {
-        return std::usize::MAX;
+        return std::u64::MAX;
     }
     1 << height_log_two
 }
 
-fn init_sized_vec<V>(bit_width: usize) -> Vec<Option<V>> {
+fn init_sized_vec<V>(bit_width: u32) -> Vec<Option<V>> {
     std::iter::repeat_with(|| None)
         .take(1 << bit_width)
         .collect()
 }
 
-fn bmap_bytes(bit_width: usize) -> usize {
+fn bmap_bytes(bit_width: u32) -> usize {
     if bit_width <= 3 {
         1
     } else {
